rename database module to room module
