--- src/server/modules/database/index.ts
@@ -1,8 +1,0 @@
-import { DatabaseModule } from "./mockDatabase";
-
-/**
- * Creates a new Logger Module for use as dependency
- */
-const DbModule = new DatabaseModule();
-
-export default DbModule;
