+++ src/server/modules/room/index.ts
@@ -1,0 +1,9 @@
+
+import { MemoryRoomModule } from "./memoryModule";
+
+/**
+ * Creates a new Logger Module for use as dependency
+ */
+const DbModule = new MemoryRoomModule();
+
+export default DbModule;
